refactor: move shortcut hint support
