+++ Sources/App/ShortcutHintSupport.swift
@@ -1,0 +1,99 @@
+import AppKit
+import Foundation
+
+enum ShortcutHintModifierPolicy {
+    static let intentionalHoldDelay: TimeInterval = 0.30
+
+    static func shouldShowHints(
+        for modifierFlags: NSEvent.ModifierFlags,
+        defaults: UserDefaults = .standard
+    ) -> Bool {
+        let normalized = modifierFlags.intersection(.deviceIndependentFlagsMask)
+            .subtracting([.numericPad, .function, .capsLock])
+        switch normalized {
+        case [.command]:
+            return ShortcutHintDebugSettings.showHintsOnCommandHoldEnabled(defaults: defaults)
+        case [.control]:
+            return ShortcutHintDebugSettings.showHintsOnControlHoldEnabled(defaults: defaults)
+        default:
+            return false
+        }
+    }
+
+    static func isCurrentWindow(
+        hostWindowNumber: Int?,
+        hostWindowIsKey: Bool,
+        eventWindowNumber: Int?,
+        keyWindowNumber: Int?
+    ) -> Bool {
+        guard let hostWindowNumber, hostWindowIsKey else { return false }
+        if let eventWindowNumber {
+            return eventWindowNumber == hostWindowNumber
+        }
+        return keyWindowNumber == hostWindowNumber
+    }
+
+    static func shouldShowHints(
+        for modifierFlags: NSEvent.ModifierFlags,
+        hostWindowNumber: Int?,
+        hostWindowIsKey: Bool,
+        eventWindowNumber: Int?,
+        keyWindowNumber: Int?,
+        defaults: UserDefaults = .standard
+    ) -> Bool {
+        shouldShowHints(for: modifierFlags, defaults: defaults) &&
+            isCurrentWindow(
+                hostWindowNumber: hostWindowNumber,
+                hostWindowIsKey: hostWindowIsKey,
+                eventWindowNumber: eventWindowNumber,
+                keyWindowNumber: keyWindowNumber
+            )
+    }
+}
+enum ShortcutHintDebugSettings {
+    static let sidebarHintXKey = "shortcutHintSidebarXOffset"
+    static let sidebarHintYKey = "shortcutHintSidebarYOffset"
+    static let titlebarHintXKey = "shortcutHintTitlebarXOffset"
+    static let titlebarHintYKey = "shortcutHintTitlebarYOffset"
+    static let paneHintXKey = "shortcutHintPaneTabXOffset"
+    static let paneHintYKey = "shortcutHintPaneTabYOffset"
+    static let alwaysShowHintsKey = "shortcutHintAlwaysShow"
+    static let showHintsOnCommandHoldKey = "shortcutHintShowOnCommandHold"
+    static let showHintsOnControlHoldKey = "shortcutHintShowOnControlHold"
+
+    static let defaultSidebarHintX = 0.0
+    static let defaultSidebarHintY = 0.0
+    static let defaultTitlebarHintX = 4.0
+    static let defaultTitlebarHintY = 0.0
+    static let defaultPaneHintX = 0.0
+    static let defaultPaneHintY = 0.0
+    static let defaultAlwaysShowHints = false
+    static let defaultShowHintsOnCommandHold = true
+    static let defaultShowHintsOnControlHold = true
+
+    static let offsetRange: ClosedRange<Double> = -20...20
+
+    static func clamped(_ value: Double) -> Double {
+        min(max(value, offsetRange.lowerBound), offsetRange.upperBound)
+    }
+
+    static func showHintsOnCommandHoldEnabled(defaults: UserDefaults = .standard) -> Bool {
+        guard defaults.object(forKey: showHintsOnCommandHoldKey) != nil else {
+            return defaultShowHintsOnCommandHold
+        }
+        return defaults.bool(forKey: showHintsOnCommandHoldKey)
+    }
+
+    static func showHintsOnControlHoldEnabled(defaults: UserDefaults = .standard) -> Bool {
+        guard defaults.object(forKey: showHintsOnControlHoldKey) != nil else {
+            return defaultShowHintsOnControlHold
+        }
+        return defaults.bool(forKey: showHintsOnControlHoldKey)
+    }
+
+    static func resetVisibilityDefaults(defaults: UserDefaults = .standard) {
+        defaults.set(defaultAlwaysShowHints, forKey: alwaysShowHintsKey)
+        defaults.set(defaultShowHintsOnCommandHold, forKey: showHintsOnCommandHoldKey)
+        defaults.set(defaultShowHintsOnControlHold, forKey: showHintsOnControlHoldKey)
+    }
+}
